dodano rutiranje i login, app prebacen u chat komponentu, zasticena app stranica s loginom, povezan username sa stateom
